+++ src/brick_games/ui/__init__.py
@@ -1,2 +1,2 @@
-from .title import TitleScreen
+from .title_screen import TitleScreen
 from .window import Window
